Restructured repo, added gym demo
