+++ src/gym_demo/gymnasium_env/__init__.py
@@ -1,0 +1,6 @@
+from gymnasium.envs.registration import register
+
+register(
+    id="gymnasium_env/GridWorld-v0",
+    entry_point="gymnasium_env.envs:GridWorldEnv",
+)
