+++ src/gym_demo/gymnasium_env/envs/__init__.py
@@ -1,0 +1,1 @@
+from gymnasium_env.envs.grid_world import GridWorldEnv
