+++ src/gym_demo/gymnasium_env/wrappers/clip_reward.py
@@ -1,0 +1,13 @@
+import gymnasium as gym
+import numpy as np
+
+
+class ClipReward(gym.RewardWrapper):
+    def __init__(self, env, min_reward, max_reward):
+        super().__init__(env)
+        self.min_reward = min_reward
+        self.max_reward = max_reward
+        self.reward_range = (min_reward, max_reward)
+
+    def reward(self, reward):
+        return np.clip(reward, self.min_reward, self.max_reward)
