+++ src/gym_demo/gymnasium_env/wrappers/discrete_actions.py
@@ -1,0 +1,12 @@
+import gymnasium as gym
+from gymnasium.spaces import Discrete
+
+
+class DiscreteActions(gym.ActionWrapper):
+    def __init__(self, env, disc_to_cont):
+        super().__init__(env)
+        self.disc_to_cont = disc_to_cont
+        self.action_space = Discrete(len(disc_to_cont))
+
+    def action(self, act):
+        return self.disc_to_cont[act]
