+++ src/gym_demo/gymnasium_env/wrappers/relative_position.py
@@ -1,0 +1,12 @@
+import gymnasium as gym
+from gymnasium.spaces import Box
+import numpy as np
+
+
+class RelativePosition(gym.ObservationWrapper):
+    def __init__(self, env):
+        super().__init__(env)
+        self.observation_space = Box(shape=(2,), low=-np.inf, high=np.inf)
+
+    def observation(self, obs):
+        return obs["target"] - obs["agent"]
